separate pipeline parallel stategies, add gpipe, not tested
