+++ symbolic_tensor_graph/graph/connect_graph.py
@@ -8,17 +8,28 @@ class ConnectGraph:
     def apply(cls, graphs, links, inplace=False, force_connect=False):
         if not inplace:
             graphs = copy.deepcopy(graphs)
-        connected_tensors = list()
+        connected_tensors = dict()
         connected_in_tensors = list()
         connected_out_tensors = list()
         connected_graph = None
+        # for tensors with same name, preserve only one, and reupdate the links
         for graph in graphs:
-            connected_tensors.extend(graph.tensors)
-            connected_in_tensors.extend(graph.in_tensors)
-            connected_out_tensors.extend(graph.out_tensors)
-            connected_graph = TensorGraph(
-                connected_tensors, connected_in_tensors, connected_out_tensors
-            )
+            for tensor in graph.tensors:
+                if tensor.id in connected_tensors:
+                    continue
+                connected_tensors[tensor.id] = tensor
+                if tensor in graph.in_tensors:
+                    connected_in_tensors.append(tensor)
+                if tensor in graph.out_tensors:
+                    connected_out_tensors.append(tensor)
+        for tensor in connected_tensors.values():
+            if tensor.x1 is not None and tensor.x1.id in connected_tensors:
+                tensor.x1 = connected_tensors[tensor.x1.id]
+            if tensor.x2 is not None and tensor.x2.id in connected_tensors:
+                tensor.x2 = connected_tensors[tensor.x2.id]
+        connected_graph = TensorGraph(
+            list(connected_tensors.values()), connected_in_tensors, connected_out_tensors
+        )
         assert connected_graph is not None
         connected_graph_id_map_graph = connected_graph.get_tensor_id_map_tensor()
         for from_, to_ in links.items():
